remove same source code, use Registry to implement Source interface. Keep the compatibility of DeployTo fields.
